+++ app/src/pages/public/components/PublicAttachmentsList.tsx
@@ -92,7 +92,7 @@ const PublicAttachmentsList: React.FC<IPublicAttachmentsListProps> = (props) => 
             <TableBody>
               {props.attachmentsList.length > 0 &&
                 props.attachmentsList.slice(page * rowsPerPage, page * rowsPerPage + rowsPerPage).map((row, index) => (
-                  <TableRow key={row.fileName}>
+                  <TableRow key={`${row.fileName}-${index}`}>
                     <TableCell scope="row">
                       <Link
                         underline="always"
